Move block rendering code to separate component; rename back Blocks.jsx to Blocks.js
